feat: unify sidebar
