--- src/components/UI/Sidebar/components/SidebarUtils/index.css
@@ -1,34 +1,0 @@
-.SidebarUtils {
-  grid-area: utils;
-  align-self: end;
-  padding-bottom: var(--sidebar-vertical-padding);
-  font-size: 16px;
-}
-
-.SidebarUtils__dropdown {
-  position: relative;
-  width: 100%;
-}
-
-.SidebarUtils__dropdownPopup {
-  position: absolute;
-  left: 100%;
-  bottom: 0;
-  width: max-content;
-  z-index: 1;
-  opacity: 0;
-  pointer-events: none;
-  transition: opacity 0.2s;
-  background: var(--navbar-active-background);
-}
-
-.SidebarUtils__dropdown:hover,
-.SidebarUtils__dropdown:focus-within {
-  background: var(--navbar-active-background);
-}
-
-.SidebarUtils__dropdown:hover .SidebarUtils__dropdownPopup,
-.SidebarUtils__dropdown:focus-within .SidebarUtils__dropdownPopup {
-  opacity: 1;
-  pointer-events: initial;
-}
